+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add optional `start_time` parameter to `SystemTracker` and `PidTracker`
 - Update `ResourceTracker` to start tracking at the nearest interval in the future, syncing `SystemTracker` and `PidTracker`
 - Fix `SystemTracker` and `PidTracker` to not drift by a few nanoseconds in every interval
+- Move cloud and server discovery to the `ResourceTracker` class from the Metaflow-specific decorators
 
 ## v0.3.1 (May 30, 2025)